requirements.txt
Django==3.2.18
django-extensions==3.1.5
PyJWT==2.4.0
djangorestframework==3.13.1
